アクティビティ（走行日）のIDをIntegerからLongに変更

+++ app/src/main/java/jp/ceed/android/mylapslogger/entity/PracticeTrack.kt
@@ -1,10 +1,9 @@
 package jp.ceed.android.mylapslogger.entity
 
 import androidx.room.ColumnInfo
-import androidx.room.Ignore
 
 data class PracticeTrack(
-    val id: Int,
+    val id  : Long,
     @ColumnInfo(name = "track_id") val trackId: Int,
     @ColumnInfo(name = "name") val trackName: String,
     @ColumnInfo(name = "lap_count") val lapCount: Int,
